Create connected version of token amount component
